pyproject.toml
[tool.poetry]
name = "api-airport-clone"
version = "0.1.0"
description = "test"
authors = ["Dmytro <94nj111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.28.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
django-debug-toolbar = "^4.4.6"
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
